Moving daemon files, renaming namespaces, modules and jar
