+++ .idea/artifacts/KotlinJpsPlugin.xml
@@ -19,8 +19,8 @@
       <element id="module-output" name="util" />
       <element id="module-output" name="util.runtime" />
       <element id="file-copy" path="$PROJECT_DIR$/resources/kotlinManifest.properties" />
-      <element id="module-output" name="kotlinr" />
-      <element id="module-output" name="rmi-interface" />
+      <element id="module-output" name="daemon-client" />
+      <element id="module-output" name="daemon-common" />
       <element id="module-output" name="deserialization" />
       <element id="extracted-dir" path="$PROJECT_DIR$/dependencies/native-platform-uberjar.jar" path-in-jar="/" />
       <element id="module-output" name="android-jps-plugin" />
