--- compiler/rmi/kotlinr/kotlinr.iml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module type="JAVA_MODULE" version="4">
-  <component name="NewModuleRootManager" inherit-compiler-output="true">
-    <exclude-output />
-    <content url="file://$MODULE_DIR$">
-      <sourceFolder url="file://$MODULE_DIR$/src" isTestSource="false" />
-    </content>
-    <orderEntry type="inheritedJdk" />
-    <orderEntry type="sourceFolder" forTests="false" />
-    <orderEntry type="library" name="kotlin-runtime" level="project" />
-    <orderEntry type="module" module-name="rmi-interface" />
-    <orderEntry type="module" module-name="frontend.java" />
-    <orderEntry type="module" module-name="util" />
-    <orderEntry type="library" name="native-platform-uberjar" level="project" />
-  </component>
-</module>
