--- compiler/rmi/rmi-server/rmi-server.iml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module type="JAVA_MODULE" version="4">
-  <component name="NewModuleRootManager" inherit-compiler-output="true">
-    <exclude-output />
-    <content url="file://$MODULE_DIR$">
-      <sourceFolder url="file://$MODULE_DIR$/src" isTestSource="false" />
-    </content>
-    <orderEntry type="inheritedJdk" />
-    <orderEntry type="sourceFolder" forTests="false" />
-    <orderEntry type="library" name="kotlin-runtime" level="project" />
-    <orderEntry type="module" module-name="cli" />
-    <orderEntry type="module" module-name="rmi-interface" />
-    <orderEntry type="module" module-name="frontend.java" />
-    <orderEntry type="module" module-name="util.runtime" />
-    <orderEntry type="module" module-name="util" />
-  </component>
-</module>
