+++ idea/idea.iml
@@ -48,7 +48,7 @@
     <orderEntry type="module" module-name="idea-completion" exported="" scope="RUNTIME" />
     <orderEntry type="module" module-name="idea-test-framework" scope="TEST" />
     <orderEntry type="module" module-name="js.serializer" />
-    <orderEntry type="module" module-name="kotlinr" />
+    <orderEntry type="module" module-name="daemon-client" />
     <orderEntry type="module" module-name="idea-repl" scope="TEST" />
     <orderEntry type="module" module-name="idea-live-templates" exported="" scope="RUNTIME" />
   </component>
